refactor: move histogram to shared components
