+++ packages/shared-components/package.json
@@ -30,7 +30,9 @@
     "tailwind-merge": "^1.8.0",
     "tailwindcss": "^3.2.4",
     "typescript": "^4.9.3",
-    "yup": "^1.0.0"
+    "yup": "^1.0.0",
+    "recharts": "^2.4.1",
+    "lodash": "^4.17.21"
   },
   "volta": {
     "extends": "../../package.json"
