+++ packages/shared-components/src/Histogram.tsx
